remove not useful test
